morphlink/requirements.txt
fastapi
uvicorn
sqlalchemy
redis
python-dotenv
httpx
